10 - Mini Auto-ML Experiment Runner for Tabular Data/auto_experiment_runner/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "autotab"
version = "0.1.0"
description = "Mini Auto-ML Experiment Runner for Tabular Data"
requires-python = ">=3.11"
dependencies = [
    "pandas",
    "numpy",
    "scikit-learn",
    "pyyaml",
    "joblib",
    "matplotlib",
    "rich",
    "xgboost",
    "pydantic"
]

[project.scripts]
autotab = "autotab.cli:main"
